Metodos Suma y Promedio Total

+++ ChallengeDiWork/Modelo/Desperfecto.cs
@@ -11,7 +11,7 @@ namespace ChallengeDiWork.Modelo
     {
         public string Descripcion { get; set; }
         public decimal ManoDeObra { get; set; }
-        public decimal Tiempo { get; set; }
+        public int Tiempo { get; set; }
         public Vehiculo Vehiculo { get; set; }
         public List<Repuesto> Repuestos { get; set; }
